+++ oshi-json/src/main/java/oshi/json/software/os/impl/FileSystemImpl.java
@@ -63,9 +63,10 @@ public class FileSystemImpl extends AbstractOshiJsonObject implements FileSystem
         oshi.software.os.OSFileStore[] fs = this.fileSystem.getFileStores();
         OSFileStore[] fileStores = new OSFileStore[fs.length];
         for (int i = 0; i < fs.length; i++) {
-            fileStores[i] = new OSFileStore(fs[i].getName(), fs[i].getVolume(), fs[i].getMount(),
+            fileStores[i] = new OSFileStore(fs[i].getName(), fs[i].getVolume(),fs[i].getMount(),
                     fs[i].getDescription(), fs[i].getType(), fs[i].getUUID(), fs[i].getUsableSpace(),
                     fs[i].getTotalSpace());
+            fileStores[i].setLogicalvolume(fs[i].getLogicalVolume());
         }
         return fileStores;
     }